requirements.txt
gradio==4.7.1
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.24.1